+++ src/s2geography/accessors-geog.h
@@ -4,6 +4,7 @@
 #include <s2/s2convex_hull_query.h>
 
 #include "s2geography/aggregator.h"
+#include "s2geography/arrow_udf/arrow_udf.h"
 #include "s2geography/geography.h"
 
 namespace s2geography {
@@ -32,5 +33,16 @@ class S2ConvexHullAggregator
   S2ConvexHullQuery query_;
   std::vector<std::unique_ptr<Geography>> keep_alive_;
 };
+
+namespace arrow_udf {
+/// \brief Instantiate an ArrowUDF for the s2_centroid() function
+///
+/// This ArrowUDF handles any GeoArrow array as input and produces
+/// a geoarrow.wkb array as output.
+std::unique_ptr<ArrowUDF> Centroid();
+std::unique_ptr<ArrowUDF> ConvexHull();
+std::unique_ptr<ArrowUDF> PointOnSurface();
+
+}  // namespace arrow_udf
 
 }  // namespace s2geography
