+++ src/s2geography/distance.cc
@@ -4,6 +4,7 @@
 #include <s2/s2closest_edge_query.h>
 #include <s2/s2furthest_edge_query.h>
 
+#include "s2geography/arrow_udf/arrow_udf_internal.h"
 #include "s2geography/geography.h"
 
 namespace s2geography {
@@ -66,5 +67,47 @@ std::pair<S2Point, S2Point> s2_minimum_clearance_line_between(
   // Find the closest point pair on edge1 and edge2.
   return S2::GetEdgePairClosestPoints(edge1.v0, edge1.v1, edge2.v0, edge2.v1);
 }
+
+namespace arrow_udf {
+
+struct S2DistanceExec {
+  using arg0_t = GeographyIndexInputView;
+  using arg1_t = GeographyIndexInputView;
+  using out_t = DoubleOutputBuilder;
+
+  void Init(const std::unordered_map<std::string, std::string>& options) {}
+
+  out_t::c_type Exec(arg0_t::c_type value0, arg1_t::c_type value1) {
+    return s2_distance(value0, value1);
+  }
+};
+
+std::unique_ptr<ArrowUDF> Distance() {
+  return std::make_unique<BinaryUDF<S2DistanceExec>>();
+}
+
+struct S2ShortestLineExec {
+  using arg0_t = GeographyIndexInputView;
+  using arg1_t = GeographyIndexInputView;
+  using out_t = WkbGeographyOutputBuilder;
+
+  void Init(const std::unordered_map<std::string, std::string>& options) {}
+
+  out_t::c_type Exec(arg0_t::c_type value0, arg0_t::c_type value1) {
+    std::pair<S2Point, S2Point> out =
+        s2_minimum_clearance_line_between(value0, value1);
+    stashed_ = PolylineGeography(std::make_unique<S2Polyline>(
+        std::vector<S2Point>{out.first, out.second}));
+    return stashed_;
+  }
+
+  PolylineGeography stashed_;
+};
+
+std::unique_ptr<ArrowUDF> ShortestLine() {
+  return std::make_unique<BinaryUDF<S2ShortestLineExec>>();
+}
+
+}  // namespace arrow_udf
 
 }  // namespace s2geography
