+++ src/s2geography/linear-referencing.cc
@@ -2,6 +2,7 @@
 #include "s2geography/linear-referencing.h"
 
 #include "s2geography/accessors.h"
+#include "s2geography/arrow_udf/arrow_udf_internal.h"
 #include "s2geography/build.h"
 #include "s2geography/geography.h"
 
@@ -72,5 +73,27 @@ S2Point s2_interpolate_normalized(const Geography& geog, double distance_norm) {
   std::unique_ptr<Geography> geog_poly = s2_rebuild(geog, GlobalOptions());
   return s2_interpolate_normalized(*geog_poly, distance_norm);
 }
+
+namespace arrow_udf {
+
+struct S2InterpolateNormalizedExec {
+  using arg0_t = GeographyInputView;
+  using arg1_t = DoubleInputView;
+  using out_t = WkbGeographyOutputBuilder;
+
+  void Init(const std::unordered_map<std::string, std::string>& options) {}
+
+  out_t::c_type Exec(arg0_t::c_type value0, arg1_t::c_type value1) {
+    stashed_ = s2_interpolate_normalized(value0, value1);
+    return stashed_;
+  }
+
+  PointGeography stashed_;
+};
+
+std::unique_ptr<ArrowUDF> InterpolateNormalized() {
+  return std::make_unique<BinaryUDF<S2InterpolateNormalizedExec>>();
+}
+}  // namespace arrow_udf
 
 }  // namespace s2geography
